tooltip utility functions moved to tooltip context
